cache fixes, ajax refresh backend
